+++ app/src/main/java/com/example/camerX/callbacks/ImageVideoResultCallback.kt
@@ -1,4 +1,4 @@
-package com.example.cameraxintegration.callbacks
+package com.example.camerX.callbacks
 
 import android.net.Uri
