+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "CameraXIntegration"
+rootProject.name = "CameraX"
 include ':app'
